requirements.txt
plotly.express
pandas
joblib